+++ src/execution/utils/uncertainNumber.ts
@@ -8,7 +8,7 @@ import type { UncertainNumber } from '../UncertainNumber';
  * @private utility for initializating UncertainNumber
  */
 export function uncertainNumber(value?: number | typeof NaN | undefined | null): UncertainNumber {
-    if (value === null || value === undefined || Number.isNaN(NaN)) {
+    if (value === null || value === undefined || Number.isNaN(value)) {
         return { value: 0, isUncertain: true };
     }