+++ src/llm-providers/_common/utils/count-total-usage/countTotalUsage.ts
@@ -42,7 +42,6 @@ export function countTotalUsage(llmTools: LlmExecutionTools): LlmExecutionToolsW
     if (llmTools.callChatModel !== undefined) {
         proxyTools.callChatModel = async (prompt: ChatPrompt): Promise<ChatPromptResult> => {
             const promptResult = await llmTools.callChatModel!(prompt);
-            console.log('!!!! promptResult.usage callChatModel', promptResult.usage);
             totalUsage = addUsage(totalUsage, promptResult.usage);
             return promptResult;
         };
@@ -51,7 +50,6 @@ export function countTotalUsage(llmTools: LlmExecutionTools): LlmExecutionToolsW
     if (llmTools.callCompletionModel !== undefined) {
         proxyTools.callCompletionModel = async (prompt: CompletionPrompt): Promise<CompletionPromptResult> => {
             const promptResult = await llmTools.callCompletionModel!(prompt);
-            console.log('!!!! promptResult.usage callCompletionModel', promptResult.usage);
             totalUsage = addUsage(totalUsage, promptResult.usage);
             return promptResult;
         };
@@ -60,7 +58,6 @@ export function countTotalUsage(llmTools: LlmExecutionTools): LlmExecutionToolsW
     if (llmTools.callEmbeddingModel !== undefined) {
         proxyTools.callEmbeddingModel = async (prompt: EmbeddingPrompt): Promise<EmbeddingPromptResult> => {
             const promptResult = await llmTools.callEmbeddingModel!(prompt);
-            console.log('!!!! promptResult.usage callEmbeddingModel', promptResult.usage);
             totalUsage = addUsage(totalUsage, promptResult.usage);
             return promptResult;
         };